requirements/requirements.txt
csr2transmart==0.0.23
docutils==0.14
lockfile==0.12.2
luigi==2.8.0
python-daemon==2.1.2
SQLAlchemy==1.3.0
tornado==4.5.2
gitpython==2.1.8
pandas>=0.25.1,<0.26.0
numpy==1.17.1
chardet==3.0.4
psycopg2-binary==2.7.4
click>=7.0,<8.0
requests>= 2.22.0,<2.23.0
